refactor: split Slack bot worker into focused modules
