+++ packages/slack-bot/src/sessions/control-plane-client.ts
@@ -1,0 +1,143 @@
+import {
+  createSessionResponseSchema,
+  sendPromptResponseSchema,
+  type CreateSessionResponse,
+  type SendPromptResponse,
+} from "@open-inspect/shared";
+import { getAuthHeaders } from "../internal-auth";
+import { createLogger } from "../logger";
+import { buildSessionTargetRequestFields, targetId, type SlackSessionTarget } from "../targets";
+import type { CallbackContext, Env } from "../types";
+
+const log = createLogger("handler");
+
+export async function createSession(
+  env: Env,
+  target: SlackSessionTarget,
+  model: string,
+  reasoningEffort: string | undefined,
+  branch: string | undefined,
+  traceId?: string,
+  slackUserId?: string,
+  actorDisplayName?: string,
+  actorEmail?: string
+): Promise<CreateSessionResponse | null> {
+  const startTime = Date.now();
+  const base = {
+    trace_id: traceId,
+    target_id: targetId(target),
+    model,
+    reasoning_effort: reasoningEffort,
+    branch,
+    slack_user_id: slackUserId,
+  };
+  try {
+    const headers = await getAuthHeaders(env, traceId);
+    const response = await env.CONTROL_PLANE.fetch("https://internal/sessions", {
+      method: "POST",
+      headers,
+      body: JSON.stringify({
+        ...buildSessionTargetRequestFields(target, branch),
+        model,
+        reasoningEffort,
+        spawnSource: "slack-bot",
+        actorUserId: slackUserId,
+        actorDisplayName,
+        actorEmail,
+      }),
+    });
+    if (!response.ok) {
+      log.error("control_plane.create_session", {
+        ...base,
+        outcome: "error",
+        http_status: response.status,
+        duration_ms: Date.now() - startTime,
+      });
+      return null;
+    }
+    const result = createSessionResponseSchema.safeParse(await response.json());
+    if (!result.success) {
+      log.error("control_plane.create_session", {
+        ...base,
+        outcome: "error",
+        error: new Error("Invalid control plane create session response"),
+        duration_ms: Date.now() - startTime,
+      });
+      return null;
+    }
+    log.info("control_plane.create_session", {
+      ...base,
+      outcome: "success",
+      session_id: result.data.sessionId,
+      http_status: 200,
+      duration_ms: Date.now() - startTime,
+    });
+    return result.data;
+  } catch (e) {
+    log.error("control_plane.create_session", {
+      ...base,
+      outcome: "error",
+      error: e instanceof Error ? e : new Error(String(e)),
+      duration_ms: Date.now() - startTime,
+    });
+    return null;
+  }
+}
+
+export async function sendPrompt(
+  env: Env,
+  sessionId: string,
+  content: string,
+  authorId: string,
+  callbackContext?: CallbackContext,
+  traceId?: string
+): Promise<SendPromptResponse | null> {
+  const startTime = Date.now();
+  const base = { trace_id: traceId, session_id: sessionId, source: "slack" };
+  try {
+    const headers = await getAuthHeaders(env, traceId);
+    const response = await env.CONTROL_PLANE.fetch(
+      `https://internal/sessions/${sessionId}/prompt`,
+      {
+        method: "POST",
+        headers,
+        body: JSON.stringify({ content, authorId, source: "slack", callbackContext }),
+      }
+    );
+    if (!response.ok) {
+      log.error("control_plane.send_prompt", {
+        ...base,
+        outcome: "error",
+        http_status: response.status,
+        duration_ms: Date.now() - startTime,
+      });
+      return null;
+    }
+    const result = sendPromptResponseSchema.safeParse(await response.json());
+    if (!result.success) {
+      log.error("control_plane.send_prompt", {
+        ...base,
+        outcome: "error",
+        error: new Error("Invalid control plane send prompt response"),
+        duration_ms: Date.now() - startTime,
+      });
+      return null;
+    }
+    log.info("control_plane.send_prompt", {
+      ...base,
+      outcome: "success",
+      message_id: result.data.messageId,
+      http_status: 200,
+      duration_ms: Date.now() - startTime,
+    });
+    return result.data;
+  } catch (e) {
+    log.error("control_plane.send_prompt", {
+      ...base,
+      outcome: "error",
+      error: e instanceof Error ? e : new Error(String(e)),
+      duration_ms: Date.now() - startTime,
+    });
+    return null;
+  }
+}
